Consolidate wwutils modules
