+++ wwutils/__init__.py
@@ -6,8 +6,8 @@ from . import misc
 from . import stats
 from . import video
 from . import dataload
+from . import plot
 from . import whisk_permissions
 from . import utils
 
-# shortcuts
-from .misc import rint, pick, pick_rows, printnow, globjoin
+# shortcutsfrom .misc import rint, pick, pick_rows, printnow, globjoin
